requirements.txt
Flask==2.2.5
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
PyJWT==2.4.0
python-dotenv==0.21.0
Werkzeug==2.2.3
python-dotenv==0.21.0
mysqlclient==2.0.3
